+++ src/pages/Attendance/MonthTable/Table.jsx
@@ -1,20 +1,17 @@
 import React, { useState, useEffect } from "react";
 import styled from "styled-components";
-import DatePicker, { registerLocale } from "react-datepicker";
 import "react-datepicker/dist/react-datepicker.css";
 import * as XLSX from "xlsx";
 import ko from "date-fns/locale/ko";
 import { GrPrevious, GrNext } from "react-icons/gr";
-import { BsCalendarDate } from "react-icons/bs"
 import { GoOctoface } from "react-icons/go"
 import { TbDog } from "react-icons/tb"
 import { RiBearSmileLine } from "react-icons/ri"
 import { AiOutlineSmile, AiOutlineDoubleLeft, AiOutlineDoubleRight } from "react-icons/ai"
 import textVariants from '../../../styles/variants/textVariants';
-import Button from "../../../components/Button";
 import Buttons from '../../../components/Buttons';
-
-registerLocale("ko", ko);
+import ClassButton from './ClassButton';
+import CustomDatepicker from '../../../components/CustomDatepicker'
 
 const Table = () => {
 
@@ -24,14 +21,15 @@ const Table = () => {
 
     const loadClassroom = (selected, id) => {
         setSelectedButton(selected);
-        // navigate(`/host/${id}/ENTER/전체시간`)
+        // navigate(`/attendance/${id}`)
     };
 
     const [selectedDate, setSelectedDate] = useState(new Date());
 
     const students = [
-        { id: 1, name: "백주원", attendanceStatus: "출석", arrivalTime: "오전 9:00", leaveTime: "오후 4:30", attendanceCnt: "3", absentCnt: "3", date: "2023-04-01" },
-        { id: 2, name: "김주원", attendanceStatus: "인정결석", arrivalTime: "오전 9:00", leaveTime: "오후 7:30", attendanceCnt: "6", absentCnt: "0", date: "2023-04-02" },
+        { id: 1, name: "백주원", attendanceStatus: "출석", enterTime: "오전 9:00", exitTime: "오후 4:30", attendanceCnt: "3", absentCnt: "3" },
+        { id: 2, name: "김주원", attendanceStatus: "인정결석", enterTime: "오전 9:00", exitTime: "오후 7:30", attendanceCnt: "6", absentCnt: "0" },
+
     ];
 
 
@@ -98,7 +96,7 @@ const Table = () => {
                 "",
                 "등원시간",
                 ...Array.from({ length: daysInMonth }, (_, i) => i + 1).map((day) => {
-                    return student.arrivalTime;
+                    return student.enterTime;
                 }),
             ];
             XLSX.utils.sheet_add_aoa(ws, [arrivalTimeData], { origin: `A${rowIndex + index * 3 + 1}` });
@@ -108,7 +106,7 @@ const Table = () => {
                 "",
                 "하원시간",
                 ...Array.from({ length: daysInMonth }, (_, i) => i + 1).map((day) => {
-                    return student.leaveTime;
+                    return student.exitTime;
                 }),
             ];
             XLSX.utils.sheet_add_aoa(ws, [leaveTimeData], { origin: `A${rowIndex + index * 3 + 2}` });
@@ -210,40 +208,14 @@ const Table = () => {
         <div>
 
             <StyledTableTitle>월별 출석부</StyledTableTitle>
-            <StyledClassButtonGroup>
-                <Button.ClassButton
-                    selected={"새빛반"}
-                    selectedButton={selectedButton}
-                    onClick={() => loadClassroom("새빛반", 1)}
-                />
-                <Button.ClassButton
-                    selected={"동동반"}
-                    selectedButton={selectedButton}
-                    onClick={() => loadClassroom("동동반", 2)}
-                />
-                <Button.ClassButton
-                    selected={"빗살반"}
-                    selectedButton={selectedButton}
-                    onClick={() => loadClassroom("빗살반", 3)}
-                />
-            </StyledClassButtonGroup>
+            <ClassButton />
             <StyledTableContainer>
                 <StyledHeader>
                     <StyledMonthYear>
                         <GrPrevious style={{ marginRight: "8px" }} onClick={decreaseMonth} size={24} />
                         {selectedDate.getFullYear()}년 {selectedDate.getMonth() + 1}월
                         <GrNext style={{ marginLeft: "8px" }} onClick={increaseMonth} size={24} />
-                        <StyledCustomDatePicker>
-                            <DatePicker
-                                selected={selectedDate}
-                                onChange={handleDateChange}
-                                customInput={<CustomInput />}
-                                dateFormat="yyyy년 MM월"
-                                locale="ko"
-                                wrapperClassName="hidden"
-                                showMonthYearPicker
-                            />
-                        </StyledCustomDatePicker>
+                        <CustomDatepicker />
                     </StyledMonthYear>
                 </StyledHeader>
                 <StyledTable>
@@ -292,13 +264,13 @@ const Table = () => {
                                 <tr>
                                     <td>등원시간</td>
                                     {visibleDays.map((day) => (
-                                        <td key={day}>{student.arrivalTime}</td>
+                                        <td key={day}>{student.enterTime}</td>
                                     ))}
                                 </tr>
                                 <tr>
                                     <td>하원시간</td>
                                     {visibleDays.map((day) => (
-                                        <td key={day}>{student.leaveTime}</td>
+                                        <td key={day}>{student.exitTime}</td>
                                     ))}
                                 </tr>
                             </>
@@ -316,9 +288,6 @@ const Table = () => {
 
 export default Table;
 
-const StyledClassButtonGroup = styled.div`
-padding-bottom: 10px;
-`;
 
 const StyledTableTitle = styled.h2`
   ${textVariants.H2_Bold}
@@ -344,17 +313,12 @@ const StyledMonthYear = styled.div`
 `;
 
 
-
-const CustomInput = React.forwardRef(({ value, onClick }, ref) => (
-    <BsCalendarDate onClick={onClick} ref={ref} className="hidden" />
-));
-
 const StyledTableContainer = styled.div`
     background-color:#EDF5EECC;
     box-shadow: 0px 2px 12px hsla(0, 0%, 0%, 0.04);
     border-radius: 12px;
     padding: 40px;
-    margin-top: 50px;
+    margin-top: 30px;
 `
 
 const StyledTable = styled.table`
@@ -411,73 +375,6 @@ const StyledButtonGroup = styled.div`
 const StyledExportButton = styled(Buttons.Filter)`
     margin-left: 10px;
 `;
-
-
-const StyledCustomDatePicker = styled.div`
-    .react-datepicker {
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    background-color: white;
-    font-size: 14px;
-  }
-  
-  .react-datepicker__header {
-    background-color: ${({ theme }) => theme.color.perple_lighter};
-    border-bottom: 1px solid #ccc;
-    font-weight: bold;
-    padding: 10px;
-    text-align: center;
-  }
-  
-  .react-datepicker__current-month {
-    ${textVariants.Body1_SemiBold}
-  }
-  
-  .react-datepicker__day {
-    ${textVariants.Body3_Regular}
-    color: ${({ theme }) => theme.color.grayScale[500]};
-    outline: none;
-  }
-  
-  .react-datepicker__day:hover {
-    background-color: ${({ theme }) => theme.color.grayScale[100]};
-    cursor: pointer;
-  }
-  
-  .react-datepicker__day--selected {
-    ${textVariants.Body3_SemiBold}
-    background-color: ${({ theme }) => theme.color.orange_lighter};
-    color: ${({ theme }) => theme.color.white};
-    
-  }
-  
-  .react-datepicker__day--today {
-    ${textVariants.Body3_SemiBold}
-    color: ${({ theme }) => theme.color.orange};
-  }
-  
-  .react-datepicker__navigation {
-    outline: none;
-    border: none;
-    background-color: transparent;
-    cursor: pointer;
-    line-height: 1.4;
-  }
-  
-  .react-datepicker__navigation--previous {
-    margin-right: 10px;
-  }
-  
-  .react-datepicker__navigation--next {
-    margin-left: 10px;
-  }
-`
-
-
-
-
-
-
